made some adjustments on Dashboard
