+++ fod_frontend/src/components/Dashboard/style.scss
