+++ fod_frontend/src/components/Location/Location.module.css
@@ -58,3 +58,14 @@
     height: 70%;
 
 }
+
+.searchButton{
+    background: #BC0000 !important;
+    border: 3px solid #BC0000 !important;
+    border-radius: 3px !important;
+}
+.searchButton:hover{
+    background: white !important;
+    color:#BC0000 !important;
+    border: 3px solid white !important;
+}
